requirements.txt
rich==12.5.1
requests
openai==0.28
pyaml
tiktoken
clickhouse-driver
